fix: decouple IdentifyDecouplingCapsSolver06 test from React page to avoid broken circuit-to-svg dependency in Bun Test
